FROMGIT media: dt-bindings: update clocks for sc7280-camss
Update clock names to make them consistent with existing platform i.e
qcom,sc8280xp-camss.yaml. Rename gcc_cam_hf_axi to gcc_axi_hf and add
gcc_axi_sf.

gcc_camera_ahb is always on and we don't need to enable it explicitly.
gcc_axi_sf is added to avoid unexpected hardware behaviour.

This change will not break ABI because the ABI hasn't been cemented yet as
the dtsi changes are not merged yet and there are no users for this driver
as of now.

Signed-off-by: Vikram Sharma <quic_vikramsa@quicinc.com>
Reviewed-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>
Reviewed-by: Krzysztof Kozlowski <krzysztof.kozlowski@linaro.org>
Signed-off-by: Hans Verkuil <hverkuil@xs4all.nl>