media: i2c: ak7375: Add support for ak7377
Add support for ak7377 VCM, which has 10-bit position values, requiring
a shift of 6 bits. It also has a different standby mode setting (0x10
instead of 0x40) and a significantly longer control delay. The maximum
focus position is 1023, and the control step count is increased to 100.

Tested on sm7325-nothing-spacewar phone.

Signed-off-by: Danila Tikhonov <danila@jiaxyga.com>