WIP usb: typec: fsa4480: Report switch & mux events to codec
In the Type-C graph, the FSA4480 switch is inbetween the USB-C connector
and the audio codec, but the audio codec also needs to receive the
switch and mux events from the FSA4480.

Update the driver to get an Type-C mux on the next endpoint and
broadcast the recevied mux & switch state to it.

Signed-off-by: Luca Weiss <luca.weiss@fairphone.com>