FROMLIST v1 Input: pm8941-pwrkey - Disable wakeup for resin by default
'Resin' (*Res*et *In*put) is usually connected to volume down buttons on
devices, which are usually not expected to wake up the device from
suspend.

On the other hand, pwrkey should keep wakeup on.

Note, that this does change behavior by turning off wakeup by default
for 'resin' and requiring a new dt property to be added to turn it on
again. But since this is not expected behavior in the first place, and
most users will not expect this, I'd argue this change is acceptable.

Signed-off-by: Luca Weiss <luca@lucaweiss.eu>